namespace refactorings
