+++ Sin.Net.Solution/Sin.Net.Persistence/Exports/JsonExporter.cs
@@ -3,6 +3,7 @@ using Sin.Net.Domain.Persistence.Adapter;
 using Sin.Net.Domain.Persistence.Logging;
 using Sin.Net.Domain.Persistence.Settings;
 using Sin.Net.Persistence.IO;
+using Sin.Net.Persistence.IO.Json;
 using Sin.Net.Persistence.Settings;
 using System.IO;
