+++ Sin.Net.Solution/Sin.Net.Persistence/IO/Json/LowerCaseContractResolver.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Sin.Net.Persistence.IO
+namespace Sin.Net.Persistence.IO.Json
 {
     internal class LowerCaseContractResolver : DefaultContractResolver
     {
